debug-engine/requirements.txt
faiss-cpu>=1.7.0
neo4j>=5.0.0
openai>=1.0.0
numpy>=1.24.0
